requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.2
django-heroku==0.3.1
django-taggit==2.1.0
Markdown==3.3.6
psycopg2==2.9.3
pyp==2.12
sqlparse==0.4.2
whitenoise==5.3.0
